qmake: change the folder structure assumed for LIBMARBLEDEVEL
LIBMARBLEDEVEL is now assumed to point to the install prefix used for a
marble install (instead of into a mix of source and build trees as it did
in the past).

This makes it much easier to have Qt4 and Qt5 versions of this library and
link to the right one as needed.

Of course this requires a matching LD_LIBRARY_PATH when launching a
such-built Subsurface.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>