Don't ignore actual diveable mixes
Fist is really obvious. I quite regularly decompress using pure O2.
It's as good a last decompression gas as you get.

The second is a bit harder. There are very few that dive with 80%+ helium
in their mixes but they exist, and there are real weirdos that dive
heliox, so they are actual diveable gases too.

Signed-off-by: Anton Lundin <glance@acc.umu.se>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>