Updates to user manual: Preferences and Main menu
This patch updates the section on Preferences to accommodate
the newer additions to this Panel.
(Three images removed, five images inserted)

The section dealing with the desciption of all the main menu items
is updated.

One small change to the FAQ section.

A few small typo edits elsewhere in manual.

Signed-off-by: Willem Ferguson <willemferguson@zoology.up.ac.za>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>